+++ src/rpc/client.cpp
@@ -176,6 +176,7 @@ static const CRPCConvertParam vRPCConvertParams[] = {
     {"buildavalancheproof", 0, "sequence"},
     {"buildavalancheproof", 1, "expiration"},
     {"buildavalancheproof", 3, "stakes"},
+    {"getstakingreward", 1, "recompute"},
 };
 
 class CRPCConvertTable {