+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.28.1 is now available from:
 
   <https://download.bitcoinabc.org/0.28.1/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - New `getstakingreward` and `setstakingreward` RPCs have been added to
+   retrieve and manually set the staking reward payout for a given block.